Extract traits
