+++ packages/forms/src/Components/Concerns/HasColors.php
@@ -1,0 +1,46 @@
+<?php
+
+namespace Filament\Forms\Components\Concerns;
+
+use Closure;
+use Illuminate\Contracts\Support\Arrayable;
+
+trait HasColors
+{
+    /**
+     * @var array<string | array{50: string, 100: string, 200: string, 300: string, 400: string, 500: string, 600: string, 700: string, 800: string, 900: string, 950: string} | null> | Arrayable | Closure
+     */
+    protected array | Arrayable | Closure $colors = [];
+
+    /**
+     * @param  array<string | array{50: string, 100: string, 200: string, 300: string, 400: string, 500: string, 600: string, 700: string, 800: string, 900: string, 950: string} | null> | Arrayable | Closure  $colors
+     */
+    public function colors(array | Arrayable | Closure $colors): static
+    {
+        $this->colors = $colors;
+
+        return $this;
+    }
+
+    /**
+     * @return string | array{50: string, 100: string, 200: string, 300: string, 400: string, 500: string, 600: string, 700: string, 800: string, 900: string, 950: string} | null
+     */
+    public function getColor(mixed $value): string | array | null
+    {
+        return $this->getColors()[$value] ?? null;
+    }
+
+    /**
+     * @return array<string | array{50: string, 100: string, 200: string, 300: string, 400: string, 500: string, 600: string, 700: string, 800: string, 900: string, 950: string} | null>
+     */
+    public function getColors(): array
+    {
+        $colors = $this->evaluate($this->colors);
+
+        if ($colors instanceof Arrayable) {
+            $colors = $colors->toArray();
+        }
+
+        return $colors;
+    }
+}
