+++ packages/forms/src/Components/Radio.php
@@ -3,16 +3,17 @@
 namespace Filament\Forms\Components;
 
 use Closure;
-use Illuminate\Contracts\Support\Arrayable;
-use Illuminate\Contracts\Support\Htmlable;
 
 class Radio extends Field implements Contracts\CanDisableOptions
 {
     use Concerns\CanDisableOptions;
     use Concerns\CanDisableOptionsWhenSelectedInSiblingRepeaterItems;
     use Concerns\CanFixIndistinctState;
+    use Concerns\HasColors;
+    use Concerns\HasDescriptions;
     use Concerns\HasExtraInputAttributes;
     use Concerns\HasGridDirection;
+    use Concerns\HasIcons;
     use Concerns\HasOptions;
 
     public const RADIOS_VIEW = 'filament-forms::components.radio.index';
@@ -27,21 +28,6 @@ class Radio extends Field implements Contracts\CanDisableOptions
     protected string $view = self::RADIOS_VIEW;
 
     protected bool | Closure $isInline = false;
-
-    /**
-     * @var array<string | array{50: string, 100: string, 200: string, 300: string, 400: string, 500: string, 600: string, 700: string, 800: string, 900: string, 950: string} | null> | Arrayable | Closure
-     */
-    protected array | Arrayable | Closure $colors = [];
-
-    /**
-     * @var array<string | Htmlable> | Arrayable | Closure
-     */
-    protected array | Arrayable | Closure $descriptions = [];
-
-    /**
-     * @var array<string | Htmlable | null> | Arrayable | Closure
-     */
-    protected array | Arrayable | Closure $icons = [];
 
     protected bool | Closure | null $isOptionDisabled = null;
 
@@ -90,120 +76,12 @@ class Radio extends Field implements Contracts\CanDisableOptions
         return $this;
     }
 
-    /**
-     * @param  array<string | array{50: string, 100: string, 200: string, 300: string, 400: string, 500: string, 600: string, 700: string, 800: string, 900: string, 950: string} | null> | Arrayable | Closure  $colors
-     */
-    public function colors(array | Arrayable | Closure $colors): static
-    {
-        $this->colors = $colors;
-
-        return $this;
-    }
-
-    /**
-     * @return string | array{50: string, 100: string, 200: string, 300: string, 400: string, 500: string, 600: string, 700: string, 800: string, 900: string, 950: string} | null
-     */
-    public function getColor(mixed $value): string | array | null
-    {
-        return $this->getColors()[$value] ?? null;
-    }
-
-    /**
-     * @return array<string | array{50: string, 100: string, 200: string, 300: string, 400: string, 500: string, 600: string, 700: string, 800: string, 900: string, 950: string} | null>
-     */
-    public function getColors(): array
-    {
-        $colors = $this->evaluate($this->colors);
-
-        if ($colors instanceof Arrayable) {
-            $colors = $colors->toArray();
-        }
-
-        return $colors;
-    }
-
-    public function disableOptionWhen(bool | Closure $callback): static
-    {
-        $this->isOptionDisabled = $callback;
-
-        return $this;
-    }
-
-    /**
-     * @param  array<string | Htmlable | null> | Arrayable | Closure  $icons
-     */
-    public function icons(array | Arrayable | Closure $icons): static
-    {
-        $this->icons = $icons;
-
-        return $this;
-    }
-
-    public function getIcon(mixed $value): string | Htmlable | null
-    {
-        return $this->getIcons()[$value] ?? null;
-    }
-
-    /**
-     * @return array<string | Htmlable | null>
-     */
-    public function getIcons(): array
-    {
-        $icons = $this->evaluate($this->icons);
-
-        if ($icons instanceof Arrayable) {
-            $icons = $icons->toArray();
-        }
-
-        return $icons;
-    }
-
     public function inline(bool | Closure $condition = true): static
     {
         $this->isInline = $condition;
         $this->inlineLabel(fn (Radio $component): ?bool => ($component->isRadios() && $component->evaluate($condition)) ? true : null);
 
         return $this;
-    }
-
-    /**
-     * @param  array<string | Htmlable> | Arrayable | Closure  $descriptions
-     */
-    public function descriptions(array | Arrayable | Closure $descriptions): static
-    {
-        $this->descriptions = $descriptions;
-
-        return $this;
-    }
-
-    /**
-     * @param  array-key  $value
-     */
-    public function hasDescription($value): bool
-    {
-        return array_key_exists($value, $this->getDescriptions());
-    }
-
-    /**
-     * @param  array-key  $value
-     */
-    public function getDescription($value): string | Htmlable | null
-    {
-        return $this->getDescriptions()[$value] ?? null;
-    }
-
-    /**
-     * @return array<string | Htmlable>
-     */
-    public function getDescriptions(): array
-    {
-        $descriptions = $this->evaluate($this->descriptions);
-
-        if ($descriptions instanceof Arrayable) {
-            $descriptions = $descriptions->toArray();
-        }
-
-        return $descriptions;
     }
 
     public function isInline(): bool
